+++ sample/gen_sample_output.sh
@@ -1,31 +1,31 @@
 #!/bin/sh
 
-cd "$(dirname $0)";
-SCRIPT_DIR="$PWD";
+cd "$(dirname $0)"
+SCRIPT_DIR="$PWD"
 
-XRESLOADER="$(ls -t ../target/xresloader-*-jar-with-dependencies.jar 2>/dev/null | head -n 1)";
+XRESLOADER="$(ls -t ../target/xresloader-*-jar-with-dependencies.jar 2>/dev/null | head -n 1)"
 if [ -z "$XRESLOADER" ] || [ ! -e "$XRESLOADER" ]; then
-    XRESLOADER="$(ls -t ../target/xresloader-*.jar 2>/dev/null | head -n 1)";
+  XRESLOADER="$(ls -t ../target/xresloader-*.jar 2>/dev/null | head -n 1)"
 fi
 
 if [ -z "$XRESLOADER" ] || [ ! -e "$XRESLOADER" ]; then
-    echo "xresloader not found.";
-    exit 0;
+  echo "xresloader not found."
+  exit 0
 fi
 
-echo "Using xresloader=$XRESLOADER";
+echo "Using xresloader=$XRESLOADER"
 
 for proto_dir in proto_v2 proto_v3; do
-    echo "Generate sample data for $proto_dir, one per cmd";
-    XLSX_FILE="$(ls *.xlsx | grep -v 'grep' | grep -v "\\~\\$")";
-    find "$proto_dir" -iname "UnreaImportSettings.json" -exec rm -fv "{}" ";" ;
-    java -client -jar "$XRESLOADER" -t lua -p protobuf -o $proto_dir -f $proto_dir/kind.pb --pretty 2 -c kind_const.lua ;
-    java -client -jar "$XRESLOADER" -t lua -p protobuf -o $proto_dir -f $proto_dir/kind.pb --pretty 2 --lua-module ProtoEnums.Kind -c kind_const_module.lua ;
-    java -client -jar "$XRESLOADER" -t bin -p protobuf -o $proto_dir -f $proto_dir/kind.pb -s $XLSX_FILE -m scheme_kind -a 1.0.0.0;
-    java -client -jar "$XRESLOADER" -t lua -p protobuf -o $proto_dir -f $proto_dir/kind.pb --pretty 4 -s $XLSX_FILE -m scheme_kind -n "/(?i)\.bin\$/\.lua/" --data-version 1.0.0.0 ;
-    java -client -jar "$XRESLOADER" -t lua -p protobuf -o $proto_dir -f $proto_dir/kind.pb --lua-module ProtoData.Kind -s $XLSX_FILE -m scheme_kind -n "/(?i)\.bin\$/_module\.lua/" --data-version 1.0.0.0 ;
-    echo "Generate sample data for $proto_dir, batchmode using --stdin";
-    CMDS='
+  echo "Generate sample data for $proto_dir, one per cmd"
+  XLSX_FILE="$(ls *.xlsx | grep -v 'grep' | grep -v "\\~\\$")"
+  find "$proto_dir" -iname "UnreaImportSettings.json" -exec rm -fv "{}" ";"
+  java -client -jar "$XRESLOADER" -t lua -p protobuf -o $proto_dir -f $proto_dir/kind.pb --pretty 2 -c kind_const.lua
+  java -client -jar "$XRESLOADER" -t lua -p protobuf -o $proto_dir -f $proto_dir/kind.pb --pretty 2 --lua-module ProtoEnums.Kind -c kind_const_module.lua
+  java -client -jar "$XRESLOADER" -t bin -p protobuf -o $proto_dir -f $proto_dir/kind.pb -s $XLSX_FILE -m scheme_kind -a 1.0.0.0
+  java -client -jar "$XRESLOADER" -t lua -p protobuf -o $proto_dir -f $proto_dir/kind.pb --pretty 4 -s $XLSX_FILE -m scheme_kind -n "/(?i)\.bin\$/\.lua/" --data-version 1.0.0.0
+  java -client -jar "$XRESLOADER" -t lua -p protobuf -o $proto_dir -f $proto_dir/kind.pb --lua-module ProtoData.Kind -s $XLSX_FILE -m scheme_kind -n "/(?i)\.bin\$/_module\.lua/" --data-version 1.0.0.0
+  echo "Generate sample data for $proto_dir, batchmode using --stdin"
+  CMDS='
         -t lua -p protobuf -o '$proto_dir'     -f '$proto_dir/kind.pb' --pretty 2 -i kind.desc.lua
         -t json -p protobuf -o '$proto_dir'    -f '$proto_dir/kind.pb' --pretty 2 -i kind.desc.json
         -t json -p protobuf -o '$proto_dir'    -f '$proto_dir/kind.pb' -s '$XLSX_FILE' -m scheme_kind -n "/(?i)\.bin$/\.json/"
@@ -37,6 +37,7 @@ for proto_dir in proto_v2 proto_v3; do
         -t lua -p protobuf -o '$proto_dir'     -f '$proto_dir/kind.pb' --pretty 2 -m DataSource='$XLSX_FILE'|arr_in_arr|3,1 -m MacroSource='$XLSX_FILE'|macro|2,1 -m ProtoName=arr_in_arr_cfg -m OutputFile=arr_in_arr_cfg.lua -m KeyRow=2
         -t bin -p protobuf -o '$proto_dir'     -f '$proto_dir/kind.pb' -m DataSource='$XLSX_FILE'|arr_in_arr|3,1 -m MacroSource='$XLSX_FILE'|macro|2,1 -m ProtoName=arr_in_arr_cfg -m OutputFile=arr_in_arr_cfg.bin -m KeyRow=2
         -t xml -p protobuf -o '$proto_dir'     -f '$proto_dir/kind.pb' --pretty 2 -m DataSource='$XLSX_FILE'|arr_in_arr|3,1 -m MacroSource='$XLSX_FILE'|macro|2,1 -m ProtoName=arr_in_arr_cfg -m OutputFile=arr_in_arr_cfg.xml -m KeyRow=2
+        -t bin -p protobuf -o '$proto_dir'    -f '$proto_dir/kind.pb' -s '$XLSX_FILE' -m scheme_upgrade --disable-excel-formular
         -t json -p protobuf -o '$proto_dir'    -f '$proto_dir/kind.pb' -s '$XLSX_FILE' -m scheme_upgrade -n "/(?i)\.bin$/\.json/" --pretty 2 --disable-excel-formular
         -t lua -p protobuf -o '$proto_dir'     -f '$proto_dir/kind.pb' -s '$XLSX_FILE' -m scheme_upgrade -n "/(?i)\.bin$/\.lua/" --pretty 2 --disable-excel-formular
         -t lua -o '$proto_dir'     -f '$proto_dir/kind.pb' --pretty 2 -i kind_option.lua
@@ -61,9 +62,9 @@ for proto_dir in proto_v2 proto_v3; do
         -t lua -p protobuf -o '$proto_dir' -f '$proto_dir/kind.pb' --pretty 2 -m DataSource='$XLSX_FILE'|keep_or_strip_empty_list|3,1 -m ProtoName=keep_or_strip_empty_list_cfg -m OutputFile=strip_list_tail_cfg.lua -m KeyRow=2 --list-strip-empty-tail
         -t ue-json -o '$proto_dir/json' -f '$proto_dir/kind.pb' -m DataSource='$XLSX_FILE'|keep_or_strip_empty_list|3,1 -m ProtoName=keep_or_strip_empty_list_cfg -m OutputFile=StripListTailCfg.json -m KeyRow=2 -m UeCfg-CodeOutput=|Public/Config|Private/Config --list-strip-empty-tail
         -t bin -p protobuf -o '$proto_dir' -f '$proto_dir/kind.pb' -m DataSource='$XLSX_FILE'|keep_or_strip_empty_list|3,1 -m ProtoName=keep_or_strip_empty_list_cfg -m OutputFile=strip_list_tail_cfg.bin -m KeyRow=2 --list-strip-empty-tail
-    ';
-    # -t ue-csv -o '$proto_dir/csv' -f '$proto_dir/kind.pb' -m DataSource='$XLSX_FILE'|keep_or_strip_empty_list|3,1 -m ProtoName=keep_or_strip_empty_list_cfg -m OutputFile=KeepEmptyListCfg.csv -m KeyRow=2 -m UeCfg-CodeOutput=|Public/Config|Private/Config --list-keep-empty
-    # -t ue-csv -o '$proto_dir/csv' -f '$proto_dir/kind.pb' -m DataSource='$XLSX_FILE'|keep_or_strip_empty_list|3,1 -m ProtoName=keep_or_strip_empty_list_cfg -m OutputFile=StripListTailCfg.csv -m KeyRow=2 -m UeCfg-CodeOutput=|Public/Config|Private/Config --list-strip-empty-tail
-    echo "Run with --stdin: $CMDS";
-    echo "$CMDS" | java -client -jar "$XRESLOADER" --stdin --data-version 1.0.0.0;
+    '
+  # -t ue-csv -o '$proto_dir/csv' -f '$proto_dir/kind.pb' -m DataSource='$XLSX_FILE'|keep_or_strip_empty_list|3,1 -m ProtoName=keep_or_strip_empty_list_cfg -m OutputFile=KeepEmptyListCfg.csv -m KeyRow=2 -m UeCfg-CodeOutput=|Public/Config|Private/Config --list-keep-empty
+  # -t ue-csv -o '$proto_dir/csv' -f '$proto_dir/kind.pb' -m DataSource='$XLSX_FILE'|keep_or_strip_empty_list|3,1 -m ProtoName=keep_or_strip_empty_list_cfg -m OutputFile=StripListTailCfg.csv -m KeyRow=2 -m UeCfg-CodeOutput=|Public/Config|Private/Config --list-strip-empty-tail
+  echo "Run with --stdin: $CMDS"
+  echo "$CMDS" | java -client -jar "$XRESLOADER" --stdin --data-version 1.0.0.0
 done